@@ -18,20 +18,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   charts_flutter: ^0.10.0
-  sqflite:
-  bloc:
-  flutter_bloc:
-  path_provider:
-  path:
-  dio:
-  socket_io_client: 
-
+  bloc: 7.0.0
+  flutter_bloc: 7.0.0
+  path_provider: ^2.0.1
+  dio: ^4.0.0
+  socket_io_client: 2.0.0-beta.4-nullsafety.0
+  firebase_auth: 1.4.1
+  firebase_core: ^1.2.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
