@@ -8,26 +8,27 @@ abstract class MonitorEvent {}
 class AskNewList extends MonitorEvent {}
 
 class ChangeForm extends MonitorEvent {
-    bool change = true;
-    ChangeForm({this.change});
+  bool change = true;
+  ChangeForm({required this.change});
 }
 
 class UpdateList extends MonitorEvent {
   List<Transaction> transactionValuesList;
   List<int> idList;
-  UpdateList({this.transactionValuesList, this.idList});
+  UpdateList({required this.transactionValuesList, required this.idList});
 }
 
 //////////////////////////////////////////////////////////////////////////
 
-
 class MonitorBloc extends Bloc<MonitorEvent, MonitorState> {
-  StreamSubscription _remoteSubscription;
+  late StreamSubscription _remoteSubscription;
 
-  List<Transaction> remoteTransactionValuesList;
-  List<int> remoteIdList;
+  late List<Transaction> remoteTransactionValuesList;
+  late List<int> remoteIdList;
 
-  MonitorBloc() : super(MonitorState(transactionValuesList: [], idList: [])) {
+  MonitorBloc()
+      : super(MonitorState(
+            transactionValuesList: [], idList: [], changeForm: true)) {
     add(AskNewList());
     _remoteSubscription = DatabaseRemoteServer.helper.stream.listen((response) {
       try {
@@ -48,20 +49,23 @@ class MonitorBloc extends Bloc<MonitorEvent, MonitorState> {
       remoteTransactionValuesList = remoteResponse[0];
       remoteIdList = remoteResponse[1];
       yield MonitorState(
+          changeForm: true, //teste
           idList: List.from(remoteIdList),
           transactionValuesList: List.from(remoteTransactionValuesList));
     } else if (event is UpdateList) {
       yield MonitorState(
+          changeForm: true, //teste
           idList: event.idList,
           transactionValuesList: event.transactionValuesList);
     } else if (event is ChangeForm) {
-        print(event.change);
-        if (event.change == true) {
-            yield MonitorState(changeTrue: event.change);
-        }
-        else if (event.change == false) {
-            yield MonitorState(changeFalse: event.change);
-        }
+      print(event.change);
+      //   if (event.change == true) {
+      //     yield MonitorState(changeTrue: event.change);
+      //   } else if (event.change == false) {
+      //     yield MonitorState(changeFalse: event.change);
+      //   }
+      yield MonitorState(
+          idList: [], transactionValuesList: [], changeForm: event.change);
     }
   }
 
@@ -71,19 +75,18 @@ class MonitorBloc extends Bloc<MonitorEvent, MonitorState> {
   }
 }
 
-
 //////////////////////////////////////////////////////////////////////////////////
-
 
 class MonitorState {
   List<Transaction> transactionValuesList;
   List<int> idList;
-  bool changeTrue;
-  bool changeFalse;
+  bool changeForm;
+//   bool changeTrue;
+//   bool changeFalse;
   MonitorState(
-      {this.transactionValuesList,
-      this.idList,
-      this.changeTrue,
-      this.changeFalse
-      });
+      {required this.transactionValuesList,
+      required this.idList,
+      //   required this.changeTrue,
+      //   required this.changeFalse
+      required this.changeForm});
 }
