@@ -1,22 +1,23 @@
 import 'package:cripose/routes.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 
-void main() {
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
   runApp(MyApp());
 }
 
 class MyApp extends StatefulWidget {
-  
-
   @override
   _MyAppState createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-    final LocalAppRouter _localAppRouter = LocalAppRouter();
+  final LocalAppRouter _localAppRouter = LocalAppRouter();
   @override
   Widget build(BuildContext context) {
-      return MaterialApp(
+    return MaterialApp(
       title: 'Flutter Demo',
       onGenerateRoute: _localAppRouter.onGenerateRoute,
     );
