@@ -9,21 +9,22 @@ class SellForm extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ManageRemoteBloc, ManageState>(builder: (context, state) {
-      if (state is InsertState) {
-        Transaction transactionValues = new Transaction();
-        return Form(
-          key: sellForm,
-          child: Column(
-            children: [
-              parmoedaFormField(transactionValues),
-              precoFormField(transactionValues),
-              qtdFormField(transactionValues),
-              submitButton(transactionValues, state, context)
-            ],
-          ),
-        );
-      }
+    return BlocBuilder<ManageRemoteBloc, ManageState>(
+        builder: (context, state) {
+      //   if (state is InsertState) {
+      Transaction transactionValues = new Transaction();
+      return Form(
+        key: sellForm,
+        child: Column(
+          children: [
+            parmoedaFormField(transactionValues),
+            precoFormField(transactionValues),
+            qtdFormField(transactionValues),
+            submitButton(transactionValues, state, context)
+          ],
+        ),
+      );
+      //   }
     });
   }
 
@@ -31,9 +32,9 @@ class SellForm extends StatelessWidget {
     return ElevatedButton(
         child: Text("Insert Data"),
         onPressed: () {
-            transactionValues.buySell = false;
-          if (sellForm.currentState.validate()) {
-            sellForm.currentState.save();
+          transactionValues.buySell = false;
+          if (sellForm.currentState!.validate()) {
+            sellForm.currentState!.save();
             BlocProvider.of<ManageRemoteBloc>(context).add(
                 SubmitTransactionEvent(transactionValues: transactionValues));
           }
@@ -45,13 +46,13 @@ class SellForm extends StatelessWidget {
       keyboardType: TextInputType.text,
       decoration: InputDecoration(labelText: "Currency pair"),
       validator: (value) {
-        if (value.length == 0) {
+        if (value!.length == 0) {
           return "Insert currency pair";
         }
         return null;
       },
       onSaved: (value) {
-        transactionValues.currencyPair = value;
+        transactionValues.currencyPair = value!;
       },
     );
   }
@@ -61,13 +62,13 @@ class SellForm extends StatelessWidget {
       keyboardType: TextInputType.number,
       decoration: InputDecoration(labelText: "Sell price"),
       validator: (value) {
-        if (double.parse(value) == 0.0) {
+        if (double.parse(value!) == 0.0) {
           return "Insert sell price";
         }
         return null;
       },
       onSaved: (value) {
-        transactionValues.orderPrice = double.parse(value);
+        transactionValues.orderPrice = double.parse(value!);
       },
     );
   }
@@ -77,13 +78,13 @@ class SellForm extends StatelessWidget {
       keyboardType: TextInputType.number,
       decoration: InputDecoration(labelText: "Quantity"),
       validator: (value) {
-        if (double.parse(value) == 0.0) {
+        if (double.parse(value!) == 0.0) {
           return "Insert quantity";
         }
         return null;
       },
       onSaved: (value) {
-        transactionValues.quantity = double.parse(value);
+        transactionValues.quantity = double.parse(value!);
       },
     );
   }
